requirements.txt
flask>=3.0.0
bcrypt>=4.0.0
gunicorn>=21.0.0
gspread>=5.12.0
google-auth>=2.23.0
APScheduler>=3.10.0
psycopg2-binary>=2.9.9
Authlib>=1.3.0
